Add trait Resource: dedup Feature and Property
